Moved title, and window size variables to the preprocessor directives

+++ Animation.cpp
@@ -20,7 +20,7 @@ void Animation::update( double elapsedTime ) {
 	this->progress += elapsedTime;	//Dodanie do obecnego stanu animacji czasu jaki uplynal od ostatniej klatki
 	double time = this->progress;	//Zmienna pomocnicza
 
-	for( int i = 0; i < this->frames.size(); i++ ) {	//Petla szuka takiego indeksu i, dla ktorego uplyniety czas odpowiada ustalonej klatce
+	for( size_t i = 0; i < this->frames.size(); i++ ) {	//Petla szuka takiego indeksu i, dla ktorego uplyniety czas odpowiada ustalonej klatce
 		time -= this->frames[i].duration;
 
 		if( time <= 0 ) { //Jesli time jest <= 0 , to znaczy ze indeks i wskazuje na klatke animacji, ktora powinna byc wyswietlona
